requirements.txt
language-check>=1.1
lxml>=4.3.2
python-docx>=0.8.10
openpyxl>=2.6.0
nltk>=3.4
googletrans>=2.4.0